requirements.txt
# litematic文件解析
litemapy>=0.7.0

# 图像处理
Pillow>=10.0.0

# 3D渲染
pyvista>=0.43.0
numpy>=1.24.0

# 数据处理
pandas>=2.0.0

# 模板引擎
Jinja2>=3.1.0

# 工具库
tqdm>=4.65.0
pyyaml>=6.0
